DEV[]MAJ: implement mutex
priority ceiling mutex is implemented
to share the led green gpio data output
registers. Now the threads can run
without the problem of priority inversion
and blocking delay can be performed.
The QF_init() function has to be called
before the BSP_init() function so that the
QXK kernel structure is ready to implement
the mutex.